Add initial support for `rgent new`
This scaffolds out a new site by creating a config in the given folder (which is created as required).
It does not add the directory structure or install themes yet. Maybe it never will. I will only use
that once, and probably don't want to maintain the added complexity from it. Plus, it makes packaging
the binary easier. The tradeoff is in the ease of starting a new project, and delivering themes.